+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Please avoid adding duplicate information across this changelog and JIRA/doc inp
   - `/db/backup` is deprecated; please use `/api/extensions/backup` instead.
   - In the GraphQL schema, the type name `Policy { whitelists: [Whitelist]! }` changes to 
     `Policy { whitelists: [Exclusion]! }` preserving the existing structure and field names.
+- Admission controller service will be deployed by default in new k8s and Openshift clusters.
+- The following  roxctl flags have been deprecated for the command `sensor generate`:
+  - `--create-admission-controller` (replaced by `--admission-controller-listen-on-creates`)
+  - `--admission-controller-enabled` (replaced by `--admission-controller-enforce-on-creates`)
 
 ## [54.0]
 - Added option to backup certificates for central.